Header cleanup of TreeDocument, c4Database, etc
Mostly just moving class declarations into standalone headers to make
them easier to find.
Also moved most functionality out of c4Database into its superclass
Database, as it makes more sense there.
Removed mustUseVersioning() method as it's unused: we only have one
form of versioning.
