+++ LiteCore/Database/Database.hh
@@ -86,6 +86,8 @@ namespace c4Internal {
         void endTransaction(bool commit);
 
         bool inTransaction() noexcept;
+        bool mustBeInTransaction(C4Error *outError) noexcept;
+        bool mustNotBeInTransaction(C4Error *outError) noexcept;
 
         class TransactionHelper {
         public:
@@ -134,6 +136,7 @@ namespace c4Internal {
         DocumentFactory& documentFactory()                  {return *_documentFactory;}
 
         fleece::impl::Encoder& sharedEncoder();
+        FLEncoder sharedFLEncoder();
 
         fleece::impl::SharedKeys* documentKeys()                  {return _dataFile->documentKeys();}
 
@@ -149,6 +152,10 @@ namespace c4Internal {
 
         BackgroundDB* backgroundDatabase();
         void stopBackgroundTasks();
+
+#if 0 // unused
+        bool mustUseVersioning(C4DocumentVersioning, C4Error*) noexcept;
+#endif
 
     public:
         // should be private, but called from Document
@@ -177,6 +184,7 @@ namespace c4Internal {
         int                         _transactionLevel {0};  // Nesting level of transaction
         unique_ptr<DocumentFactory> _documentFactory;       // Instantiates C4Documents
         unique_ptr<fleece::impl::Encoder> _encoder;         // Shared Fleece Encoder
+        FLEncoder                   _flEncoder {nullptr};   // Ditto, for clients
         unique_ptr<access_lock<SequenceTracker>> _sequenceTracker; // Doc change tracker/notifier
         mutable unique_ptr<BlobStore> _blobStore;           // Blob storage
         uint32_t                    _maxRevTreeDepth {0};   // Max revision-tree depth
@@ -185,69 +193,5 @@ namespace c4Internal {
         Retained<Housekeeper>       _housekeeper;           // for expiration/cleanup tasks
     };
 
-
-    static inline C4Database* external(Database *db)    {return (C4Database*)db;}
-
-
-    /** Abstract interface for creating Document instances; owned by a Database. */
-    class DocumentFactory {
-    public:
-        DocumentFactory(Database *db)       :_db(db) { }
-        Database* database() const          {return _db;}
-
-        virtual ~DocumentFactory() { }
-        virtual Retained<Document> newDocumentInstance(C4Slice docID) =0;
-        virtual Retained<Document> newDocumentInstance(const Record&) =0;
-        virtual Retained<Document> newLeafDocumentInstance(C4Slice docID, C4Slice revID, bool withBody) =0;
-
-        virtual alloc_slice revIDFromVersion(slice version) =0;
-        virtual bool isFirstGenRevID(slice revID)               {return false;}
-
-    private:
-        Database* const _db;
-    };
-
-
-    /** DocumentFactory subclass for rev-tree document schema. */
-    class TreeDocumentFactory : public DocumentFactory {
-    public:
-        TreeDocumentFactory(Database *db)   :DocumentFactory(db) { }
-        Retained<Document> newDocumentInstance(C4Slice docID) override;
-        Retained<Document> newDocumentInstance(const Record&) override;
-        Retained<Document> newLeafDocumentInstance(C4Slice docID, C4Slice revID, bool withBody) override;
-        alloc_slice revIDFromVersion(slice version) override;
-        bool isFirstGenRevID(slice revID) override;
-        static slice fleeceAccessor(slice docBody);
-        
-        static Document* documentContaining(const fleece::impl::Value *value) {
-            auto doc = treeDocumentContaining(value);
-            return doc ? doc : leafDocumentContaining(value);
-        }
-
-    private:
-        static Document* treeDocumentContaining(const fleece::impl::Value *value);
-        static Document* leafDocumentContaining(const fleece::impl::Value *value);
-    };
-
 }
-
-
-// This is the struct that's forward-declared in the public c4Database.h
-struct c4Database : public c4Internal::Database {
-    c4Database(const FilePath &path, C4DatabaseConfig config)
-    :Database(path, config) { }
-    
-    ~c4Database();
-    
-    FLEncoder sharedFLEncoder();
-
-    bool mustUseVersioning(C4DocumentVersioning, C4Error*) noexcept;
-    bool mustBeInTransaction(C4Error *outError) noexcept;
-    bool mustNotBeInTransaction(C4Error *outError) noexcept;
-
-    C4ExtraInfo extraInfo { };
-
-private:
-    FLEncoder                   _flEncoder {nullptr};
-};
