+++ LiteCore/Database/Document.hh
@@ -39,8 +39,6 @@ namespace fleece { namespace impl {
 } }
 
 namespace c4Internal {
-    class TreeDocument;
-
 
     /** A versioned LiteCore document.
         This is an abstract base class whose concrete subclasses are TreeDocument (rev-trees) 
@@ -65,12 +63,13 @@ namespace c4Internal {
         // Returns a new Document object identical to this one (doesn't copy the doc in the db!)
         virtual Document* copy() =0;
 
+#if 0 // unused
         bool mustUseVersioning(C4DocumentVersioning requiredVersioning, C4Error *outError) {
-            return external(_db)->mustUseVersioning(requiredVersioning, outError);
+            return _db->mustUseVersioning(requiredVersioning, outError);
         }
-
+#endif
         bool mustBeInTransaction(C4Error *outError) {
-            return external(_db)->mustBeInTransaction(outError);
+            return _db->mustBeInTransaction(outError);
         }
 
         Database* database()    {return _db;}
@@ -214,5 +213,24 @@ namespace c4Internal {
     static inline Document *asInternal(C4Document *doc) {
         return (Document*)doc;
     }
+
+
+    /** Abstract interface for creating Document instances; owned by a Database. */
+    class DocumentFactory {
+    public:
+        DocumentFactory(Database *db)       :_db(db) { }
+        Database* database() const          {return _db;}
+
+        virtual ~DocumentFactory() { }
+        virtual Retained<Document> newDocumentInstance(C4Slice docID) =0;
+        virtual Retained<Document> newDocumentInstance(const Record&) =0;
+        virtual Retained<Document> newLeafDocumentInstance(C4Slice docID, C4Slice revID, bool withBody) =0;
+
+        virtual alloc_slice revIDFromVersion(slice version) =0;
+        virtual bool isFirstGenRevID(slice revID)               {return false;}
+
+    private:
+        Database* const _db;
+    };
 
 } // end namespace
